refactor: 알림 공용 훅 위치 정리
